Fixed errors caused by dynamic variable cast
During compilation, Make complained that in initializer list cannot be any
implicitly casted values so I've fixed this issue.

+++ common/viewer.cpp
@@ -1691,8 +1691,8 @@ namespace rs2
         auto x = static_cast<float>(-M_PI / 2);
         float _rx[4][4] = {
             { 1 , 0, 0, 0 },
-            { 0, cos(x), -sin(x), 0 },
-            { 0, sin(x), cos(x), 0 },
+            { 0, static_cast<float>(cos(x)), static_cast<float>(-sin(x)), 0 },
+            { 0, static_cast<float>(sin(x)), static_cast<float>(cos(x)), 0 },
             { 0, 0, 0, 1 }
         };
         auto z = static_cast<float>(M_PI);
